requirements.txt
anyio==3.3.4
argon2-cffi==21.1.0
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
colored==1.4.3
cycler==0.10.0
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
einops==0.3.2
entrypoints==0.3
fonttools==4.27.1
Gooey==1.0.8.1
idna==3.3
imageio==2.9.0
imageio-ffmpeg==0.4.5
imutils==0.5.4
iniconfig==1.1.1
ipykernel==6.4.2
ipympl==0.8.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==8.0.0a6
jedi==0.18.0
Jinja2==3.0.2
json5==0.9.6
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.9.0rc0
jupyter-server==1.11.1
jupyterlab==4.0.0a14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==2.0.0a3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0rc1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
moviepy==2.0.0.dev2
natsort==7.1.1
nbclassic==0.3.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nd2reader==3.3.0
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.5
numpy==1.21.2
olefile==0.46
opencv-python==4.5.4.58
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
PIMS==0.5
pip==21.3.1
pluggy==1.0.0
proglog==0.1.9
prometheus-client==0.11.0
prompt-toolkit==3.0.21
psutil==5.8.0
py==1.10.0
pycparser==2.20
Pygments==2.10.0
pygtrie==2.4.2
pyparsing==3.0.1
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
pywin32==302
pywinpty==1.1.4
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-unixsocket==0.2.0
scikit-image==0.18.3
scipy==1.7.1
Send2Trash==1.8.1b0
setuptools==58.0.4
setuptools-scm==6.3.2
six==1.16.0
slicerator==1.0.0
sniffio==1.2.0
terminado==0.12.1
testpath==0.5.0
tifffile==2021.10.12
toml==0.10.2
tomli==1.2.2
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
wheel==0.37.0
widgetsnbextension==4.0.0a4
wincertstore==0.2
wxPython==4.1.1
xmltodict==0.12.0
